requirements.txt
# 核心依赖
crawl4ai>=0.6.3
google-genai>=1.12.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0

# 辅助依赖
requests>=2.31.0
configparser>=5.3.0

# 日志和工具
colorama>=0.4.6
tqdm>=4.66.0

# HTML处理
lxml>=4.9.0
html5lib>=1.1

# 异步支持
aiohttp>=3.8.0
asyncio-throttle>=1.0.0


playwright>=1.40.0
